Here's the most significant missing piece of functionality from the mime display code:  The ability to do something with a message/rfc822 other than look at it and toggle it's header display.
This patch adds a command to the mouse-right menu for message/rfc822 parts of
"Extract Message".  Selecting this pulls the message out of the enclosing
message and puts it into the current folder.

From there, you can do anything to this message that you can do with any other
message.