Added a dropKeys pattern to pgpGPG.tcl.
Simplified PGP GUI by removing "detached" signature option. Problem
was that MIME+standard includes a copy of the message being signed
in the signature attachement. What really should be used is
MIME+detached. Decided to overload the meaning of "standard". If
the format is plain, standard means "binary". If the format is
anything else, standard means "detached". Less flexibility but
better chances of "doing the right thing".